[tests][s]: moved push related tests into subdirectory so we don't run them on every commit (we'll run them on tagged commits) - refs #291
